heartbeat for event stream
This adds a heartbeat check for the event stream.  A heartbeat ticker
runs a container on a specified interval and watches for the create
event.  When the create event is detected an in memory key will be
updated.  If no event is received, the key expires and the heartbeat
check will warn and trigger a reconnect.

Signed-off-by: Evan Hazlett <ejhazlett@gmail.com>